move methods to map.cpp and mapgen.cpp
