+++ initialize_map.cpp
@@ -7,6 +7,8 @@
 #include "draw.hpp"
 #include "elona.hpp"
 #include "item.hpp"
+#include "map.hpp"
+#include "mapgen.hpp"
 #include "variables.hpp"
