+++ main.hpp
@@ -16,8 +16,6 @@ extern int prm_818;
 extern int prm_820;
 extern int prm_853;
 
-extern int tmp_at_m172;
-
 extern elona_vector1<int> inhlist_at_m184;
 extern elona_vector1<int> inhlist;
