+++ variables.hpp
@@ -26,6 +26,21 @@ ELONA_EXTERN(elona_vector1<int> ap);
 ELONA_EXTERN(int wishfilter);
 ELONA_EXTERN(int nooracle);
 
+// map.cpp/mapgen.cpp
+ELONA_EXTERN(int tile_trap);
+ELONA_EXTERN(int tile_doorclosed4);
+ELONA_EXTERN(int tile_dooropen);
+ELONA_EXTERN(int tile_downlocked);
+ELONA_EXTERN(int tile_re);
+ELONA_EXTERN(int tile_plant);
+ELONA_EXTERN(int tile_board);
+ELONA_EXTERN(int tile_votebox);
+ELONA_EXTERN(int tile_townboard);
+ELONA_EXTERN(elona_vector1<int> isetgod);
+ELONA_EXTERN(elona_vector1<int> isetcrop);
+ELONA_EXTERN(elona_vector1<int> fsetwear);
+ELONA_EXTERN(int rdtry);
+
 ELONA_EXTERN(elona_vector1<int> _randcolor);
 ELONA_EXTERN(elona_vector1<int> aniref);
 ELONA_EXTERN(elona_vector1<int> anix);
@@ -590,7 +605,6 @@ int customtalk(int = 0, int = 0);
 int direction(int = 0, int = 0, int = 0, int = 0);
 int discsetmc();
 int dist(int = 0, int = 0, int = 0, int = 0);
-int dist_town();
 int dmgcon(int = 0, int = 0, int = 0);
 int dmghp(int = 0, int = 0, int = 0, int = 0, int = 0);
 void dmgmp(int, int);
@@ -650,13 +664,6 @@ int try_to_disarm_trap();
 int try_to_perceive_npc(int);
 int do_create_character();
 int do_create_item(int, int, int);
-int initialize_quest_map_crop();
-int initialize_random_nefia_rdtype1();
-int initialize_random_nefia_rdtype4();
-int initialize_random_nefia_rdtype5();
-int initialize_random_nefia_rdtype2();
-int initialize_random_nefia_rdtype3();
-int initialize_quest_map_party();
 int initialize_world_map();
 int label_1753();
 int get_card_info();
@@ -1383,21 +1390,6 @@ void lovemiracle(int = 0);
 void make_dish(int = 0, int = 0);
 void make_sound(int = 0, int = 0, int = 0, int = 0, int = 0, int = 0);
 void makecardlist();
-void map_converttile();
-void map_createroomdoor();
-void map_initcustom(const std::string&);
-void map_initialize();
-void map_makedoor();
-void map_nextdir1(int = 0, int = 0);
-void map_nextdir2(int = 0, int = 0);
-void map_placearena(int = 0, int = 0);
-void map_placecharaonentrance(int = 0, int = 0, int = 0);
-void map_placeplayer();
-void map_randomtile(int = 0, int = 0);
-void map_randsite(int = 0, int = 0);
-void map_reload(const std::string&);
-void map_setfog(int = 0, int = 0);
-void map_tileset(int = 0);
 void mapitem_cold(int = 0, int = 0);
 void mapitem_fire(int = 0, int = 0);
 void matdelmain(int = 0, int = 0);
